@@ -1,40 +1,54 @@
 {
   "name": "chop_chow_sd",
   "version": "0.1.0",
+  "engines": {
+    "node": "16.x"
+  },
   "private": true,
-  "proxy": "http://localhost:5000",
   "dependencies": {
-    "@material-ui/core": "^4.10.2",
-    "@material-ui/lab": "^4.0.0-alpha.55",
-    "bootstrap": "^4.3.1",
+    "@aws-sdk/client-cognito-identity": "^3.41.0",
+    "@aws-sdk/client-s3": "^3.41.0",
+    "@aws-sdk/credential-provider-cognito-identity": "^3.41.0",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/lab": "^5.0.0-alpha.86",
+    "@mui/material": "^5.8.4",
+    "@mui/styles": "^5.8.4",
+    "axios": "^0.27.2",
+    "connected-react-router": "^6.8.0",
+    "ejs": "^3.1.6",
     "jquery": "^3.5.1",
-    "material-ui-chip-input": "^1.1.0",
-    "mdbreact": "^4.27.0",
-    "node-sass": "^4.14.1",
+    "os": "^0.1.2",
     "passport-facebook": "^3.0.0",
+    "path": "^0.12.7",
     "pure-react-carousel": "^1.25.1",
-    "react": "^16.8.6",
+    "react": "17.0.2",
     "react-bootstrap": "^1.0.0-beta.8",
-    "react-bootstrap-typeahead": "^3.4.6",
-    "react-dom": "^16.8.6",
-    "react-page-layout": "^0.9.8",
+    "react-bootstrap-typeahead": "^6.0.0-rc.1",
+    "react-dom": "17.0.2",
+    "react-redux": "^7.2.1",
     "react-responsive-carousel": "^3.1.50",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.1",
-    "reactjs-popup": "^1.5.0",
-    "source-map-explorer": "^1.8.0"
+    "react-scripts": "^2.1.3",
+    "react-shadow-scroll": "^2.0.0",
+    "react-to-print": "^2.14.4",
+    "reactjs-popup": "^2.0.5",
+    "redux": "^4.0.5",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0",
+    "source-map-explorer": "^2.5.2"
   },
   "scripts": {
     "analyze": "source-map-explorer 'build/static/js/*.js'",
+    "start-js": "react-scripts start",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "storybook": "start-storybook -p 9009 -s public",
-    "build-storybook": "build-storybook -s public"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
+    "build-storybook": "build-storybook -s public",
+    "build-css": "node-sass src/ -o src/"
   },
   "browserslist": [
     ">0.2%",
@@ -44,10 +58,7 @@
   ],
   "devDependencies": {
     "@babel/core": "^7.5.5",
-    "@storybook/addon-actions": "^5.1.10",
-    "@storybook/addon-links": "^5.1.10",
-    "@storybook/addons": "^5.1.10",
-    "@storybook/react": "^5.1.10"
+    "npm-run-all": "^4.1.3"
   },
   "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
   "main": "index.js",
