@@ -1,13 +1,18 @@
 import React from "react";
 import "./style.css";
 import PageTitle from "../CommonComponents/PageTitle";
-import { Spinner } from "react-bootstrap";
 import { Container, Alert, Card, Col, Row, Button } from "react-bootstrap";
 import { Typeahead } from "react-bootstrap-typeahead";
 import ProductDetail from './ProductDetail/ProductModal'
+import { connect } from 'react-redux';
+import { withRouter } from "react-router-dom";
+import axios from '../../util/Api';
+import 'react-bootstrap-typeahead/css/Typeahead.css';
+
 
 //////////////////////////////////////////////////////////////////////
-export default class GroceryPage extends React.Component {
+class GroceryPage extends React.Component {
+
   // Mongo
   _isMounted = false;
   products = [];
@@ -58,66 +63,54 @@ export default class GroceryPage extends React.Component {
       this.setState({ product_img: _img, product_name: _name , product_modal_flg:true, productID: productID});
     }
   }
+
   onCloseClicked = () => {
     this.setState({ product_modal_flg: false });
   }
+
 //////////////////////////////////////////////////////////////////////
   componentDidMount() {
     this._isMounted = true;
+
     if (this._isMounted) {
-      const { auth, customerId } = this.props;
-      this.setState({ Authentication: auth });
-      this.setState({ customerId: customerId });
-      this.getCustomerList(customerId);
+      const { authUser, customer_id } = this.props;
+      this.setState({ Authentication: authUser });
+      this.setState({ customerId: customer_id });
+    }
+    if(this.props.customerId){
+      this.getCustomerList(this.props.customer_id);
+    }else if(localStorage.getItem("customerList")){
+      this.setState({customerList: JSON.parse(localStorage.getItem("customerList"))})
     }
   }
 
   //////////////////////////////////////////////////////////////////////
   componentWillReceiveProps(nextProps) {
     // checks if user is already logged in in app.
-    const { auth, customerId } = nextProps;
-    console.log("comes in grocery page cdm");
-    this.setState({ Authentication: auth });
-    this.setState({ customerId: customerId });
+    const { authUser, customer_id } = nextProps;
+    this.setState({ Authentication: authUser });
 
-    console.log("this.props, ", nextProps);
-
-    if (auth === true) {
-      // or if (customerId !== null) , grocery page not displaying after login click
-      this.getCustomerList(customerId);
+    if (authUser !== null) {
+      this.setState({ customerId: customer_id });
+      this.getCustomerList(customer_id);
+    }else if(localStorage.getItem("customerList")){
+      this.setState({customerList: JSON.parse(localStorage.getItem("customerList"))})
     }
   }
 
   //////////////////////////////////////////////////////////////////////
   getCustomerList = (customerId) => {
-    var localToken = window.localStorage.getItem("userToken");
-    console.log("customder id  iss: " + customerId);
-    var url = `./api/getCustomerGroceryList/${customerId}`;
+    var url = `/getCustomerGroceryList/${customerId}`;
     // var url = `http://localhost:5000/api/getCustomerGroceryList/${customerId}`
-
-    fetch(url, {
-      method: "GET",
-      // credentials: 'include',
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: "Bearer " + localToken,
-      },
-    })
-      .then((res) => {
-        console.log("customer list response is ");
-        console.log(res);
-        return res.json();
-      })
-      .then((response) => {
-        if (response) {
-          this.setState({ customerList: response.data });
-        }
-      })
+    // var url = `https://chopchowdev.herokuapp.com/api/getCustomerGroceryList/${customerId}`;
+    axios(url)
+      .then(({data}) => {
+        this.setState({ customerList: data.data });
+      })      
       .catch(() => {
         this.setState(
           {
-            messageAlert:
-              "Authentication Error while fetching your grocery list...",
+            messageAlert:"Authentication Error while fetching your grocery list...",
             showAlert: true,
             variant: "danger",
           },
@@ -126,52 +119,16 @@ export default class GroceryPage extends React.Component {
               this.setState({ messageAlert: "", showAlert: false });
             }, 8000)
         );
-      });
-
-    // // url = "https://chopchowdev.herokuapp.com/api/get-all-products";
-    // // url = `http://localhost:5000/api/get-all-products`
-    // url = "./api/get-all-products";
-    // // or should we call this in App.js and pass it as a prop ??
-
-    // fetch(url, {
-    //   method: "GET",
-    //   // credentials: 'include',
-    //   // headers: {
-    //   //   'Content-Type': 'application/json',
-    //   // }
-    // })
-    //   .then((res) => res.text())
-    //   .then((body) => {
-    //     // console.log("should print body");
-    //     // console.log(body);
-    //     var productsList = JSON.parse(body);
-    //     console.log("PRINTING ALL PRODUCTS LIST");
-    //     // console.log(productsList);
-    //     if (productsList && productsList.data.length !== 0) {
-    //       console.log("returns GET ALL PRODUCTS ");
-    //       console.log(productsList.data.length);
-    //       for (var i = 0; i < productsList.data.length; i++) {
-    //         this.products.push(productsList.data[i]);
-    //         this.productNamesForTypeahead.set(
-    //           productsList.data[i].product_name,
-    //           productsList.data[i].id
-    //         );
-    //       }
-    //       console.log(this.products);
-    //       console.log(this.productNamesForTypeahead);
-    //     } else {
-    //       console.log("get all products function does not return");
-    //     }
-    //   })
-    //   .catch((err) => {
-    //     console.log(err);
-    //   });
+      });    
   };
 
   //////////////////////////////////////////////////////////////////////
-  handleShowDeleteItem = (productID) => {
-    this.setState({ deletedItemId: productID });
+  handleShowDeleteItem = (product_name, productID) => {
     const { customerId, deletedItemId } = this.state;
+    
+    if(this.state.customerId !== undefined && productID !== -1){
+      this.setState({ deletedItemId: productID });
+
     // var url = `https://chopchowdev.herokuapp.com/api/remove-item/${productID}/${customerId}`
     var url = `./api/remove-item/${productID}/${customerId}`;
 
@@ -203,9 +160,11 @@ export default class GroceryPage extends React.Component {
             // do we need catch sttmnt for filter
             (item) => item.id !== deletedItemId
           );
+          console.log(newValueData)
+          localStorage.setItem("customerList", JSON.stringify(newValueData))
           return { customerList: newValueData };
         });
-        console.log("Delets item");
+        console.log("Deletse item");
         this.componentDidMount();
       })
       .catch(() => {
@@ -221,12 +180,40 @@ export default class GroceryPage extends React.Component {
             }, 8000)
         );
       });
+    }
+    else{
+
+      // Remove from displyed list with images
+      this.setState((prevState) => {
+        // delete item on client side
+        const newValueData = prevState.customerList.filter(
+          // do we need catch sttmnt for filter
+          (item) => item.product_name !== product_name
+        );
+        console.log(newValueData)
+        localStorage.setItem("customerList", JSON.stringify(newValueData))
+        return { customerList: newValueData };
+      });
+      // remove from customer List variable
+      let temp_list = this.state.customerList; // make a separate copy of the array
+      let index = temp_list.indexOf(product_name);
+      if (index !== -1) {
+        temp_list.splice(index, 1);
+        this.setState({ customerList: JSON.stringify(temp_list) });
+        localStorage.setItem("customerList", temp_list)
+      }
+      // remove from selected list state
+
+      // selected = selected.filter(e => e !== product_name);
+    }
   };
 
   //////////////////////////////////////////////////////////////////////
   handleDeleteList = () => {
     console.log("Comes in deletes list");
     const { customerId } = this.state;
+    if(customerId !== undefined){
+
     var url = `https://chopchowdev.herokuapp.com/api/remove-list/${customerId}`;
     // var url = `./api/remove-list/${customerId}`
 
@@ -253,6 +240,7 @@ export default class GroceryPage extends React.Component {
         );
 
         this.setState({ customerList: [] });
+        localStorage.removeItem("customerList")
         this.componentDidMount();
         console.log("deletes list");
         return response.json();
@@ -271,67 +259,134 @@ export default class GroceryPage extends React.Component {
             }, 8000)
         );
       });
+            
+    }
+    else{
+      localStorage.removeItem("customerList")
+      this.setState({customerList: []})
+    }
   };
 
   //////////////////////////////////////////////////////////////////////
   handleClickTypeahead = (selected) => {
+    // var multiSelections = this.props.productNames
+    // const [multiSelections , setMultiSelections] = useState([]);
+
+    // var arrayOfProductNames = Array.from(this.props.productNames);
+    var foundIndex = this.props.productNames.findIndex((el) => el === selected[0]);
     this.setState({ selectedProduct: selected });
-    // var arrayOfProductNames = Array.from(this.productNamesForTypeahead.keys());
-
-    console.log("selected is:");
-    console.log(selected[0]);
-    console.log(typeof selected[0]);
-    console.log("array of product names is:");
-    console.log(this.productNamesForTypeahead);
-
     this.productNamesForTypeahead.get(selected);
-
-    // var index = arrayOfProductNames.findIndex((el) => el === selected[0]);
     var productID = this.productNamesForTypeahead.get(selected[0]);
     console.log("productID is: " + productID);
-    console.log("customer id is: " + this.state.customerId);
-    if (!isNaN(productID)) {
+    // console.log("customer id is: " + this.state.customerId);
+    console.log(selected);
+    let selectedProductsListLength = selected.length;
+
+    if(selectedProductsListLength === 0){
+      console.log("sees that selected is empty");
+      return;
+    }
+    if (!isNaN(productID) && this.state.customerId != null) {
+      console.log("Comes in if");
       // var url = `https://chopchowdev.herokuapp.com/api/addTypeaheadDataToCustomerGroceryList/${productID}/${this.state.customerId}`
       var url = `./api/addTypeaheadDataToCustomerGroceryList/${productID}/${this.state.customerId}`;
       fetch(url, {
         method: "POST",
       }).then((response) => {
         if (response) {
-          this.setState(
-            {
-              messageAlert: "product added successfully",
-              showAlert: true,
-              variant: "success",
-            },
-            () =>
-              setTimeout(() => {
-                this.setState({ messageAlert: "", showAlert: false });
-              }, 3500)
-          );
+          
           console.log("Comes in handleClickTypeahead's then on client side");
           this.componentDidMount();
         }
       });
     }
+    if(foundIndex !== -1){
+      var mealObject = {
+        id : foundIndex,
+        product_name : selected[selectedProductsListLength-1],
+        product_image: 'chopchow_default_instruction.png',
+        size: 'N/A'
+      };
+
+      console.log(mealObject);
+    }
+    else{
+      console.log("Comes in else, means it soes not have a product id for this product");
+      console.log(foundIndex);
+      console.log(selected);
+
+      // add to local list
+      // var mealObject = Object;
+      mealObject = {
+        id : -1,
+        product_name : selected[0].label,
+        product_image: 'butter.jpg',
+        size: 'N/A'
+      };
+    }
+
+    if(this.state.customerList == null){
+      this.setState({customerList: [mealObject]})
+      localStorage.setItem("customerList", JSON.stringify([mealObject]))
+    }
+    else{
+      let temp_list = this.state.customerList
+      temp_list.push(mealObject);
+      localStorage.setItem("customerList", JSON.stringify(temp_list))
+      this.setState({customerList: temp_list})
+    }
+
+    this.setState(
+      {
+        messageAlert: "product added successfully",
+        showAlert: true,
+        variant: "success",
+      },
+      () =>
+        setTimeout(() => {
+          this.setState({ messageAlert: "", showAlert: false });
+        }, 3500)
+    );
+    selected = [];
+    return;
   };
 
   //////////////////////////////////////////////////////////////////////
   render() {
     const { showAlert, variant, messageAlert, customerList } = this.state;
+    //create product names to always exclude what is already in the customers list
+    let productNamesMinusCustomersList = this.props.productNames;
+    if(customerList!= null){
+      for(let i = 0 ; i < this.state.customerList.length; i++){
+        let curr_product_name = customerList[i].product_name;
+        productNamesMinusCustomersList = productNamesMinusCustomersList.filter(e => e !== curr_product_name);
+      }
+    }
+    const ref = React.createRef();
+
+
     console.log("Authentication, ", this.state.Authentication);
     console.log("customerId, ", this.state.customerId);
     console.log("customerList, ", this.state.customerList);
 
+    console.log("Props:", this.props)
+  
+
     return (
       <>
         <Typeahead
-          // multiple
-          options={Array.from(this.productNamesForTypeahead.keys())}
+          allowNew
+          multiple
+          options={productNamesMinusCustomersList}
           placeholder="Add products to your grocery list here.."
           id="typeahead"
           onChange={(selected) => {
             this.handleClickTypeahead(selected);
+            ref.current.clear();
           }}
+          // create reference to clear after onChange
+          ref={ref}
+          // selected={this.props.productNames}
         />
 
         {/* Display alert if there is any issue loading grocery page */}
@@ -339,53 +394,50 @@ export default class GroceryPage extends React.Component {
           {messageAlert}
         </Alert>
 
+
+        <h2>Your Grocery Cart</h2>
+
+
         {this.state.messageVisible ? (
           <div>
             you can not add in this item because it is already in customers
             grocery list
           </div>
         ) : null}
-        {this.state.Authentication ? (
           <>
-            <PageTitle title=" Your Grocery List" />
+          <PageTitle title=" Your Grocery List" />
             <div>
               <Button
                   className="yourlist__buttonDeleteList"
                   variant="danger"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    this.handleDeleteList();
-                  }}
-                  >
+                  onClick={(e) => { e.stopPropagation(); this.handleDeleteList();}}>
                   Delete List Items
               </Button>
 
               <br></br>
               <Container className="page__container" fluid>
                 {/* display grocery list, for any authenticated customer */}
-                {customerList ? (
+                {
+                customerList ? (
                   customerList.map((customer_grocery_product_item) => {
+                    // console.log(customerList);
+                    console.log(customer_grocery_product_item);
                     let productID = customer_grocery_product_item.id;
-
+                    console.log("customer_grocery_product_item:", customer_grocery_product_item.product_image)
                     return (
                       // <>
-                      <Row
-                        display="inline-flex"
-                        key = {customer_grocery_product_item.id}
-                      >
-                        <Col key={customer_grocery_product_item.id}>
+                      <Row display="inline-flex" key = {customer_grocery_product_item.product_name} >
+                        <Col key={customer_grocery_product_item.product_name+'images'}>
                           {/* check for private or public images (can be used for suggest meal) */}
-                          {customer_grocery_product_item.product_image.startsWith(
-                            "http://"
-                          ) ||
-                            customer_grocery_product_item.product_image.startsWith(
-                              "data"
-                            ) ? (
+                          {customer_grocery_product_item.product_image.startsWith('https://') ? (
                               <img
-                                src={`${customer_grocery_product_item.product_image}`}
+                                src={customer_grocery_product_item.product_image}
                                 alt="product_img "
                                 className="card-img"
-                                onClick = {() => this.handleProductClick(customer_grocery_product_item.product_image, customer_grocery_product_item.product_name, productID, false)}
+                                onClick = {() => this.handleProductClick(
+                                  customer_grocery_product_item.product_image,
+                                   customer_grocery_product_item.product_name, productID,
+                                    false)}
                               />
                             ) : (
                               <img
@@ -397,38 +449,31 @@ export default class GroceryPage extends React.Component {
                             )}
                         </Col>
 
-                        <Col>
+                        <Col key={customer_grocery_product_item.product_name+'details'}>
                           <Card.Title className="grocery_item_card-header">
                             Product Name :{" "}
                             {customer_grocery_product_item.product_name}
                           </Card.Title>
                           <Card.Text>
                             Product Price :{" "}
-                            {customer_grocery_product_item &&
-                              customer_grocery_product_item.product_price}
+                            {/* {customer_grocery_product_item && */}
+                            {customer_grocery_product_item.product_price}
                             <br></br>
-                            Product Size : {customer_grocery_product_item.sizes}
+                            Product Size : {customer_grocery_product_item.size}
                           </Card.Text>
                         </Col>
 
                         <Col>
-                          <Button
-                            onClick={(e) => {
-                              e.stopPropagation();
-                              this.handleAddItemToCart(productID);
-                            }}
-                          >
-                            {" "}
-                            Add To Cart
+                          <Button onClick={(e) => { e.stopPropagation(); this.handleAddItemToCart(productID);  }} >
+                            {" "}  Add To Cart
                           </Button>
                         </Col>
                         <Col>
                           <i
                             className="fa fa-remove"
-                            onClick={(e) => {
-                              e.stopPropagation();
-                              this.handleShowDeleteItem(
-                                customer_grocery_product_item.id
+                            onClick={(e) => {e.stopPropagation();this.handleShowDeleteItem(
+                                customer_grocery_product_item.product_name,
+                                customer_grocery_product_item.productID,
                               );
                             }}
                           ></i>
@@ -437,21 +482,15 @@ export default class GroceryPage extends React.Component {
                     );
                   })
                 ) : (
-                    <Spinner animation="border" variant="info" />
+                  // consider getting items from their cache
+                  //  to display something when users arent logged in
+                    <div> Your grocery cart looks empty.
+                      {/* <Spinner animation="border" variant="info" /> */}
+                    </div>
                   )}
               </Container>
             </div>
           </>
-        ) : (
-            <>
-              {/* <Login /> */}
-              <div>
-                Log into your account or continue as guest to load your grocery
-                list
-            </div>
-            </>
-          )}
-
         {
           this.state.product_modal_flg && 
           <ProductDetail state = { this.state } onCloseClicked={this.onCloseClicked} />
@@ -460,3 +499,9 @@ export default class GroceryPage extends React.Component {
     );
   }
 }
+const mapStateToProps = ({ auth }) => {
+  const { authUser, role, customer_id } = auth;
+  return { authUser, role, customer_id }
+};
+
+export default connect(mapStateToProps, ()=>({}))(withRouter(GroceryPage));
