@@ -20,7 +20,7 @@ class RecipeContentSection extends Component {
                     </div>
                     <div>{this.props.selectedMeal.readTime} </div>
                     <div> {this.props.selectedMeal.cookTime}</div>
-                    <Popup trigger={<u>Clickon here to read steps</u>} modal 
+                    <Popup trigger={<u>Click here to read steps</u>} modal 
                         contentStyle={contentStyle}>
                         {close => (
                             <div>
