@@ -1,21 +1,61 @@
 import React, { Component } from 'react';
 import "react-responsive-carousel/lib/styles/carousel.min.css";
 import { Carousel } from 'react-responsive-carousel';
- 
+import no_mealImg from '../../../assets/images/no_meal_step_image.png';
+// import "./WithScrollbar.scss";
+
 export default class WithScrollbar extends Component {
+  constructor() {
+    super();
 
-    render() {
-        return (
-            <Carousel showThumbs={false} infiniteLoop={true} centerMode={true} centerSlidePercentage={100 / 2}>
-              {this.props.products.map(ingredient => (
-                <div key={ingredient}>
-                    <p className="">{ingredient.ingredient}</p>
-                    <img src={"/images/products/"+ingredient.image} alt={ingredient.ingredient} />
+    // this.state = {  width: 0 };
+    // window.addEventListener("resize", this.update);
+  }
 
-                </div>
-              ))}
-               
-            </Carousel>
-        );
-    }
+  // update = () => {
+  //   this.setState({  width: window.innerWidth });
+  //   let count = 1;
+  //   if (this.state.width > 1200) count = 4;
+  //   else if(this.state.width > 1000 && this.state.width < 1200) count = 3;
+  //   else if(this.state.width > 800 && this.state.width < 1000) count = 2;
+
+  //   if(this.props.products === null && this.state.width > 800) count = 4;
+  //   else if(this.props.products.length < 4 && this.state.width > 800) count = 4; //Math.min(count, this.props.products.length);
+  // };
+
+  render() {
+    let col_count = this.props.col_count;
+    console.log(this.props.productsObj);
+
+    return (
+      <div>
+  
+
+        Gets here value
+      
+        <Carousel 
+        className="withScroll_slider" 
+        showThumbs={false} 
+        infiniteLoop={false} 
+        selectedItem = { parseInt(this.props.products.length /2) } 
+        centerMode={true} 
+        centerSlidePercentage={100 / (1.4*col_count)}
+        >
+          {this.props.products.map((ingredient, index) => (        
+            <div key={index}>                    
+                <img src={ingredient.productImgPath!=="" && 
+                this.props.productsObj[ingredient.productName] != undefined ?
+                 this.props.productsObj[ingredient.productName]["image"] : no_mealImg} 
+                 alt={ingredient.ingredient} 
+                 height="200px" 
+                 width="auto"/>
+                <p style={{fontSize:"20px"}}>
+                  {ingredient.productName}</p>
+            </div>
+          ))}
+            
+        </Carousel>
+        </div>
+    );
+  }
 }
